refactor: only transpile for browser build
We can run the original sources with Node, this simplifies the package
use in Node as a consumer doesn't have to deal with source maps to get
the correct exception lines. You don't have to deal with transpiling the
source each time they make a change.
