+++ .eslintignore
@@ -1,2 +1,1 @@
 dist/*.js
-lib/
