+++ .gitignore
@@ -58,4 +58,3 @@ typings/
 .next
 
 dist/
-lib/
