+++ lib/array-slice.js
