+++ lib/minim.js
