+++ lib/namespace.js
