+++ lib/object-slice.js
