+++ lib/primitives/element.js
