+++ lib/primitives/null-element.js
