+++ lib/primitives/string-element.js
