+++ package.json
@@ -6,13 +6,13 @@
   "scripts": {
     "prepare": "npm run build",
     "build": "npm run build:cjs && npm run build:browser",
-    "build:cjs": "babel src --out-dir lib",
-    "build:browser": "browserify -d -s minim -t [ babelify ] -o dist/minim.js src/minim.js",
+    "build:cjs": "babel lib --out-dir dist/lib",
+    "build:browser": "browserify -d -s minim -t [ babelify ] -o dist/minim.js dist/lib/minim.js",
     "coverage": "istanbul cover _mocha -- -R spec --recursive --require babel-register",
     "coveralls": "coveralls <coverage/lcov.info",
     "frontendtests": "karma start",
     "lint": "eslint .",
-    "nodejstests": "mocha -R spec --recursive --require babel-register",
+    "nodejstests": "mocha -R spec --recursive",
     "precoveralls": "npm run coverage",
     "prepublish": "npm run lint",
     "test": "npm run nodejstests && npm run frontendtests"
