+++ test/converters-test.js
@@ -1,5 +1,5 @@
 const { expect } = require('./spec-helper');
-const minim = require('../src/minim').namespace();
+const minim = require('../lib/minim').namespace();
 
 describe('Minim Converters', () => {
   describe('convertToElement', () => {
