+++ test/elements/link-element-test.js
@@ -1,5 +1,5 @@
 const { expect } = require('../spec-helper');
-const minim = require('../../src/minim').namespace();
+const minim = require('../../lib/minim').namespace();
 
 const LinkElement = minim.getElementClass('link');
