+++ test/primitives/member-element-test.js
@@ -1,5 +1,5 @@
 const { expect } = require('../spec-helper');
-const minim = require('../../src/minim').namespace();
+const minim = require('../../lib/minim').namespace();
 
 const MemberElement = minim.getElementClass('member');
