+++ test/subclass-test.js
@@ -1,5 +1,5 @@
 const { expect } = require('./spec-helper');
-const minim = require('../src/minim').namespace();
+const minim = require('../lib/minim').namespace();
 
 const ArrayElement = minim.getElementClass('array');
 const StringElement = minim.getElementClass('string');
